point taken vladdy
